Frontend code review
